requirements.txt
numpy==1.23.1
Pillow==9.2.0
tensorboardX==2.5.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0

